[Instrumentation.AWS] Move adding request and response tags to AWSTracingPipelineHandler
