+++ src/OpenTelemetry.Instrumentation.AWS/Implementation/AWSTracingPipelineCustomizer.cs
@@ -7,8 +7,8 @@ using Amazon.Runtime.Internal;
 namespace OpenTelemetry.Instrumentation.AWS.Implementation;
 
 /// <summary>
-/// Wires <see cref="AWSPropagatorPipelineHandler"/> into the AWS <see cref="RuntimePipeline"/>
-/// so it can inject trace headers and add request information to the tags.
+/// Wires <see cref="AWSTracingPipelineHandler"/> and <see cref="AWSPropagatorPipelineHandler"/> into the AWS
+/// <see cref="RuntimePipeline"/> so they can inject trace headers and add request information to the tags.
 /// </summary>
 internal class AWSTracingPipelineCustomizer : IRuntimePipelineCustomizer
 {
@@ -34,7 +34,12 @@ internal class AWSTracingPipelineCustomizer : IRuntimePipelineCustomizer
             return;
         }
 
-        var propagatingPipelineHandler = new AWSPropagatorPipelineHandler(this.options);
+        var tracingPipelineHandler = new AWSTracingPipelineHandler(this.options);
+        var propagatingPipelineHandler = new AWSPropagatorPipelineHandler();
+
+        // AWSTracingPipelineHandler must execute early in the AWS SDK pipeline
+        // in order to manipulate outgoing requests objects before they are marshalled (ie serialized).
+        pipeline.AddHandlerBefore<Marshaller>(tracingPipelineHandler);
 
         // AWSPropagatorPipelineHandler executes after the AWS SDK has marshalled (ie serialized)
         // the outgoing request object so that it can work with the request's Headers
